feat: Migrate service authentication and access attributes.
Instead of using Enums to define them, we now have their own classes.

Users will need to tweak the `utils/initialize_database.py` to do the
initial load of these elements, and then manually edit in the database
the elements for their ids, so for example changing `"access":
"internet"` to `"access": "net_001"`.

doc: Add the ADR for the risk report